src/services/etl/requirements.txt
pandas==2.0.2
pyspark==3.4.0
python-decouple==3.8
redis==4.5.5
requests==2.31.0
